requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.9
pytest==8.4.2
httpx==0.26.1
PyYAML==6.0.2